ctf_solver/requirements.txt
paramiko
pyyaml
google-generativeai